group all the pre-render functionalities into separate mods
