+++ src/tests/mod.rs
@@ -1,4 +1,4 @@
+mod escape_test;
+
 #[cfg(feature = "pre-render")]
 mod render_test;
-
-mod escaping_test;
